requirements.txt
keboola.component==1.2.0
logging_gelf==0.0.18
mock
freezegun
pandas
azure-storage-blob==12.3.1
azure-identity==1.3.1
azure-storage-file-datalake
fnmatch2